Put common tasks on script js
