+++ detail.html
@@ -77,7 +77,7 @@
                     </li>
                     <li class="nav-item mx-lg-5 mx-md-3 mx-sm-0">
                       <a class="nav-link  " id="pills-Ingredient-tab" data-toggle="pill" href="#pills-Ingredient" role="tab"
-                        aria-controls="pills-Ingredient" aria-selected="false">Ingrediant</a>
+                        aria-controls="pills-Ingredient" aria-selected="false">Ingredient</a>
                     </li>
                     <li class="nav-item mx-lg-5 mx-md-3 mx-sm-0">
                       <a class="nav-link " id="pills-instraction-tab" data-toggle="pill" href="#pills-instraction" role="tab"
@@ -137,39 +137,11 @@
                             
                         </div>
                       </template>
-                      <div class="row round mb-5">
-                        <div class="col-md-4 mx-0 px-0 ">
-                          <img class="img-fluid roundImg h-100  mx-0 w-100" src="./assets/images/asset-4.png" alt="">
-                        </div>
-                        <div class="col-md-8 pt-3 pl-5 ">
-                          <h1 class="h1">
-                            Name of Ingrediant
-                          </h1>
-                          <p >
-
-                            Aisle : Pasta and Rice <br>
-                            Made up of : about frozen cauliflower florets, thawed, cut into bite-sized pieces <br>
-                            Amount : 445.54 ml
-                          </p>
-
-                        </div>
-
-                      </div>
                     </div>
   <!--======================================================********============********============********=====================================================-->
                     <div class="tab-pane fade " id="pills-instraction" role="tabpanel" aria-labelledby="pills-instraction-tab">
 
                       <p class="text-justify mb-5 instructionP" >
-                        Lorem ipsum dolor sit amet, consectetur adipisicing elit. Ipsa expedita optio, ea corrupti excepturi
-                        accusamus exercitationem doloribus distinctio inventore consequatur quam necessitatibus voluptatem
-                        pariatur laborum, aperiam autem architecto? Explicabo, eaque!
-                      </p>
-                      <p class="text-justify instructionP" >
-                        Lorem ipsum dolor sit amet, consectetur adipisicing elit. Ipsa expedita optio, ea corrupti excepturi
-                        accusamus exercitationem doloribus distinctio inventore consequatur quam necessitatibus voluptatem
-                        pariatur laborum, aperiam autem architecto? Explicabo, eaque!
-                      </p>
-                      <p class="text-justify instructionP" >
                         Lorem ipsum dolor sit amet, consectetur adipisicing elit. Ipsa expedita optio, ea corrupti excepturi
                         accusamus exercitationem doloribus distinctio inventore consequatur quam necessitatibus voluptatem
                         pariatur laborum, aperiam autem architecto? Explicabo, eaque!
@@ -214,16 +186,14 @@
                             <tr>
 
                               <td>5</td>
-                              <td class="text-center">Ketogenic</td>
+                              <td class="text-center">Price per Serving</td>
                               <td class="text-center">true/false</td>
                             </tr>
-                            <tr>
                             <tr>
 
                               <td>6</td>
                               <td class="text-center">cheap</td>
                               <td class="text-center">true/false</td>
-                            </tr>
                             </tr>
                           </tbody>
                         </table>
@@ -287,6 +257,7 @@
   </script>
   <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
   <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.bundle.min.js"></script>
+  <script type="module" src="./assets/js/script.js"></script>
   <script type="module" src="./assets/js/detail.js"></script>
 </body>
