+++ index.html
@@ -73,7 +73,7 @@
                           <h5 class="card-title">Pakistani Chicken platter</h5>
                           <p class="card-text">This dishe is famous in pakistan and made for special events.</p>
                         </div>
-                        <a href="#" class="btn">See full Recipe</a>
+                        <a href="#" class="btn" id="detail-link">See full Recipe</a>
                       </div>
                     </div>
                   </div>
@@ -143,5 +143,6 @@
             <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.bundle.min.js"></script>
             <!-- type = module is required to use imports  -->
             <script type="module" src="./assets/js/script.js"></script>
+            <script type="module" src="./assets/js/home.js"></script>
         </body>
     </html>
